+++ internal/nocalhost-api/sleep/constants.go
@@ -1,0 +1,13 @@
+package sleep
+
+const (
+	KActive      = "active"
+	KAsleep      = "asleep"
+	KDeployment  = "Deployment"
+	KStatefulSet = "StatefulSet"
+	KConfig      = "nocalhost.dev.sleep/config"
+	KStatus      = "nocalhost.dev.sleep/status"
+	KReplicas    = "nocalhost.dev.sleep/replicas"
+	KForceSleep  = "nocalhost.dev.sleep/force-sleep"
+	KForceWakeup = "nocalhost.dev.sleep/force-wakeup"
+)
